+++ apps/index.html
@@ -2,6 +2,7 @@
 <html lang="ja">
 <head>
     <script src="https://code.jquery.com/jquery.min.js"></script>
+    <script src="index.js"></script>
 <script>
 $(function () {
                 $(".example4").click(function () {
@@ -101,14 +102,13 @@ $(function () {
         </div>
         
         <p>
-            <button type="button" value="START &#13;&#10;SPEAKING" class="example4" onclick="onBtnClick();" id="button">
+            <button type="button" value="START &#13;&#10;SPEAKING" class="example4" onclick="speak();" id="button">
                 START<br>SPEAKING
             </button>
         </p>
     </div>
 </section>
-<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-<script type="module" src="index.js"> </script>
-<script src="test.js"> </script>
+
+<script src="index.js"> </script>
 </body>
 </html>
